updated column cfee datatype

+++ TechBoxtoUnbox/Models/DTO/Course.cs
@@ -9,7 +9,7 @@ namespace TechBoxtoUnbox.Models.DTO
         
         public string CName { get; set; }
         
-        public decimal CFee { get; set; }
+        public double CFee { get; set; }
         public DateTime Createddate { get; set; }
     }
 }
